Create mainNormalSycl.cpp
This is not the final version. 
I have still some problem regarding the programming (how to use two devices in SYCL) but mainly with the machine that I am testing SYCL with: It seems that I cannot find neither the CPU or the GPU. I think that I will need to try to install SYCL again.